# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentación de mi API con Echo.",
        "title": "API de Sliceflow",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1000",
    "basePath": "/",
    "paths": {
...
        "/hornero/auth/login": {
            "post": {
                "description": "Autentica al usuario y devuelve un token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authenticator"
                ],
                "summary": "Iniciar sesión",
                "parameters": [
                    {
                        "description": "Credenciales",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.UserLoginCreds"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.TokenResponse"
                        }
                    }
                }
            }
        },
...
        "/hornero/authed/machine/addmac": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Machines"
                ],
                "summary": "Crear nueva impresora 3D",
                "parameters": [
                    {
                        "description": "Datos Máquina",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.CreateMachineDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/hornero/authed/machine/delmac/{id}": {
            "delete": {
                "description": "Saca la máquina del listado de disponibles.",
                "tags": [
                    "Machines"
                ],
                "summary": "Eliminar una impresora (Borrado lógico)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la Máquina",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/hornero/authed/machine/updmac/{id}": {
            "put": {
                "tags": [
                    "Machines"
                ],
                "summary": "Actualizar impresora",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Máquina",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos Nuevos",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.UpdateMachineDTO"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/hornero/authed/materials/addmat": {
            "post": {
                "description": "Registra un insumo (Filamento, Resina, etc) para usar en las órdenes de producción.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Production"
                ],
                "summary": "Crear un nuevo material",
                "parameters": [
                    {
                        "description": "Datos del Material",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.CreateMaterialDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/hornero/authed/materials/delmat/{id}": {
            "delete": {
                "description": "Marca un insumo como eliminado. No lo borra físicamente de la DB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Production"
                ],
                "summary": "Eliminar un material (Borrado Lógico)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del Material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/hornero/authed/materials/list": {
            "get": {
                "description": "Obtiene la lista de insumos disponibles (excluye eliminados).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Production"
                ],
                "summary": "Listar todos los materiales activos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.Material"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/hornero/authed/materials/updmat/{id}": {
            "put": {
                "description": "Modifica los datos de un insumo por su ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Production"
                ],
                "summary": "Actualizar un material existente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del Material",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevos datos",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_francotraversa_Sliceflow_internal_types.UpdateMaterialDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_francotraversa_Sliceflow_internal_types.CreateMachineDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Ej: \"Prusa i3 MK3S+\"",
                    "type": "string"
                },
                "type": {
                    "description": "Ej: \"FDM\", \"SLS\"",
                    "type": "string"
                }
            }
        },
        "github_com_francotraversa_Sliceflow_internal_types.CreateMaterialDTO": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_francotraversa_Sliceflow_internal_types.Material": {
            "type": "object",
            "properties": {
                "brand": {
                    "description": "Ej: \"Marca Grilon, temperatura 200°C\"",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "Ej: \"Marca Grilon, temperatura 200°C\"",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "Ej: \"PLA Negro\"",
                    "type": "string"
                },
                "type": {
                    "description": "Ej: \"Filamento\", \"Resina\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_francotraversa_Sliceflow_internal_types.TokenResponse": {
            "type": "object",
            "properties": {
                "expires": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "github_com_francotraversa_Sliceflow_internal_types.UpdateMachineDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "\"idle\", \"printing\", \"maintenance\"",
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_francotraversa_Sliceflow_internal_types.UpdateMaterialDTO": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_francotraversa_Sliceflow_internal_types.UserLoginCreds": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "hornero3dx"
                },
                "username": {
                    "type": "string",
                    "example": "hornero3dx"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribí \"Bearer \" seguido de tu token JWT. Ejemplo: \"Bearer eyJhbG...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}